created more functions in kpi_utils to be used

+++ packet_0xB8D8_processor.py
@@ -43,7 +43,7 @@ class Packet_0xB8D8:
                         new_value = abs(number)
                         modified_entry[new_key] = new_value
                     elif new_key == 'SINR(dBm)':
-                        new_value = int(value)
+                        new_value = Decimal(value)
                         # new_value = Decimal(value)
                         modified_entry[new_key] = new_value
                     else:
